Bumping packages
and moving tests into their own files. One function per file.
