+++ package.json
@@ -24,7 +24,6 @@
     "buffer": "^6.0.3",
     "d3": "^7.8.2",
     "element-plus": "^2.2.28",
-    "pg": "^8.11.2",
     "pinia": "^2.0.29",
     "prettier": "^3.0.0",
     "vue": "^3.3.4",
@@ -56,8 +55,8 @@
     "eslint-plugin-vue": "^9.9.0",
     "happy-dom": "^10.0.3",
     "jsdom": "^22.1.0",
-    "msw": "^1.2.1",
-    "typescript": "^5.1.6",
+    "msw": "^1.2.5",
+    "typescript": "^5.2.2",
     "vite": "^4.4.9",
     "vite-plugin-mkcert": "^1.13.3",
     "vite-tsconfig-paths": "^4.2.0",
